Cleanup integration tests
