+++ docs/general/02_building-and-testing.md
@@ -18,12 +18,11 @@ The framework can be unit-tested by `cd`ing into `test` folder and calling
 dotnet msbuild /t:RunTests
 ```
 
-Moreover, there are some integration-style tests that require external services. They can be tested with `docker` and `docker-compose` tool. There are two integration-test suites:
+Moreover, there are some integration-style tests that require external services. They can be tested with `docker` and `docker-compose` tool. There is one integration-test suite currently:
 
- 1. `test/IntegrationTests` - _real_ integration tests,
- 2. `test/Testing/LeanCode.IntegrationTestHelpers.Tests` - test for the integration test helpers.
+ 1. `test/LeanCode.IntegrationTests` - _real_ integration tests,
 
-Both have `docker` folder that contains necessary configuration. You can run each of the suites using:
+It has a `docker` folder that contains necessary configuration. You can run the suite using:
 
 ```sh
 docker-compose run test
