--- test/IntegrationTests/LeanCode.IntegrationTests.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../../src/Core/LeanCode.Components/LeanCode.Components.csproj" />
-    <ProjectReference Include="../../src/Core/LeanCode.Components.Startup/LeanCode.Components.Startup.csproj" />
-    <ProjectReference Include="../../src/Core/LeanCode.Pipelines.Autofac/LeanCode.Pipelines.Autofac.csproj" />
-    <ProjectReference Include="../../src/Domain/LeanCode.IdentityProvider/LeanCode.IdentityProvider.csproj" />
-    <ProjectReference Include="../../src/Domain/LeanCode.TimeProvider/LeanCode.TimeProvider.csproj" />
-    <ProjectReference Include="../../src/Infrastructure/LeanCode.Firebase/LeanCode.Firebase.csproj" />
-    <ProjectReference Include="../../src/Infrastructure/LeanCode.Firebase.FCM/LeanCode.Firebase.FCM.csproj" />
-    <ProjectReference Include="../../src/Testing/LeanCode.IntegrationTestHelpers/LeanCode.IntegrationTestHelpers.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-
-    <PackageReference Include="Microsoft.EntityFrameworkCore" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
-  </ItemGroup>
-
-</Project>
