--- test/IntegrationTests/docker/.env
@@ -1,1 +1,0 @@
-COMPOSE_PROJECT_NAME=corelib-integration-tests
