+++ test/LeanCode.IntegrationTests/docker/Docker.Directory.Build.props
