+++ test/Testing/LeanCode.IntegrationTestHelpers.Tests/App/ApiModule.cs
@@ -1,7 +1,7 @@
-using System.IdentityModel.Tokens.Jwt;
+using System;
 using Autofac;
 using LeanCode.Components;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -19,41 +19,16 @@ namespace LeanCode.IntegrationTestHelpers.Tests.App
 
         public override void ConfigureServices(IServiceCollection services)
         {
-            services.AddMemoryCache();
-
-            services.AddIdentityServer()
-                .AddInMemoryApiResources(AuthConfig.GetApiResources())
-                .AddInMemoryIdentityResources(AuthConfig.GetIdentityResources())
-                .AddInMemoryApiScopes(AuthConfig.GetApiScopes())
-                .AddInMemoryClients(AuthConfig.GetClients())
-                .AddTestUsers(AuthConfig.GetUsers())
-                .AddDeveloperSigningCredential();
-
-            JwtSecurityTokenHandler.DefaultInboundClaimTypeMap.Clear();
-
-            services.AddAuthentication(
-                options =>
-                {
-                    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(cfg =>
-                {
-                    cfg.Authority = UrlHelper.Concat(config.GetValue<string>("InternalBase"), "auth");
-                    cfg.TokenValidationParameters.ValidateAudience = false;
-                    cfg.TokenValidationParameters.ValidateIssuer = false;
-                    cfg.RequireHttpsMetadata = false;
-
-                    cfg.TokenValidationParameters.RoleClaimType = "role";
-                });
-
-            services.AddDbContext<TestDbContext>(cfg =>
-                cfg.UseSqlServer(config.GetConnectionString("Database")));
+            var connStr = new SqliteConnectionStringBuilder
+            {
+                DataSource = Guid.NewGuid().ToString("N"),
+                Mode = SqliteOpenMode.Memory,
+                Cache = SqliteCacheMode.Shared,
+            };
+            services.AddDbContext<TestDbContext>(cfg => cfg.UseSqlite(connStr.ConnectionString));
         }
 
         protected override void Load(ContainerBuilder builder)
-        {
-            builder.RegisterType<AppRoles>().AsImplementedInterfaces();
-        }
+        { }
     }
 }
