+++ test/Testing/LeanCode.IntegrationTestHelpers.Tests/App/Startup.cs
@@ -1,12 +1,7 @@
-using LeanCode.Cache.AspNet;
 using LeanCode.Components;
 using LeanCode.Components.Startup;
-using LeanCode.CQRS.Cache;
 using LeanCode.CQRS.Default;
 using LeanCode.CQRS.RemoteHttp.Server;
-using LeanCode.CQRS.Security;
-using LeanCode.CQRS.Validation;
-using LeanCode.CQRS.Validation.Fluent;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 
@@ -14,7 +9,7 @@ namespace LeanCode.IntegrationTestHelpers.Tests.App
 {
     public class Startup : LeanStartup
     {
-        private static readonly TypesCatalog CQRSTypes = TypesCatalog.Of<ApiModule>();
+        private static readonly TypesCatalog CQRSTypes = TypesCatalog.Of<Startup>();
 
         protected override IAppModule[] Modules { get; }
 
@@ -23,26 +18,15 @@ namespace LeanCode.IntegrationTestHelpers.Tests.App
         {
             Modules = new IAppModule[]
             {
-                new CQRSModule()
-                    .WithCustomPipelines<AppContext>(
-                        CQRSTypes,
-                        c => c.Secure().Validate(),
-                        q => q.Secure().Cache()),
-
-                new InMemoryCacheModule(),
-                new FluentValidationModule(CQRSTypes),
-
                 new ApiModule(config),
+                new CQRSModule()
+                    .WithCustomPipelines<Context>(CQRSTypes, c => c, q => q),
             };
         }
 
         protected override void ConfigureApp(IApplicationBuilder app)
         {
-            app
-                .Map("/auth", inner => inner.UseIdentityServer())
-                .Map("/api", cfg => cfg
-                    .UseAuthentication()
-                    .UseRemoteCQRS(CQRSTypes, AppContext.FromHttp));
+            app.Map("/api", x => x.UseRemoteCQRS(CQRSTypes, c => new Context()));
         }
     }
 }
