--- test/Testing/LeanCode.IntegrationTestHelpers.Tests/docker/docker-compose.yml
@@ -1,38 +1,0 @@
-version: "3.6"
-services:
-  #### Tests
-  test:
-    build:
-      context: ../../../..
-      dockerfile: ./test/Testing/LeanCode.IntegrationTestHelpers.Tests/docker/Dockerfile
-    volumes:
-      - ../../../..:/app/code
-      - ~/.nuget:/home/dotnet/.nuget
-    environment:
-      - ConnectionStrings__DatabaseBase=Server=db,1433;User Id=sa;Password=Passw12#
-    depends_on:
-      - db
-
-  watch_test:
-    build:
-      context: ../../../..
-      dockerfile: ./test/Testing/LeanCode.IntegrationTestHelpers.Tests/docker/Dockerfile.watch
-    volumes:
-      - ../../../..:/app/code
-      - ~/.nuget:/home/dotnet/.nuget
-    environment:
-      - ConnectionStrings__DatabaseBase=Server=db,1433;User Id=sa;Password=Passw12#
-    depends_on:
-      - db
-
-  #### Infrastructure
-  db:
-    image: microsoft/mssql-server-linux:latest
-    environment:
-      - ACCEPT_EULA=Y
-      - SA_PASSWORD=Passw12#
-    volumes:
-      - database:/var/opt/mssql
-
-volumes:
-  database:
